+++ docs/whats_new.rst
@@ -16,6 +16,8 @@ New Features
 
 - Add `Query.__repr__`
 - Let :func:`flogin.search_handler.SearchHandler.callback` and :func:`flogin.search_handler.SearchHandler.context_menu` return ``None``
+- Add ``Generic`` to :class:`flogin.search_handler.SearchHandler` for :attr:`flogin.search_handler.SearchHandler.plugin`
+- Add ``Generic`` to :class:`flogin.jsonrpc.results.Result` for :attr:`flogin.jsonrpc.results.Result.plugin`
 
 v0.0.5
 -------